+++ src/core/spy.ts
@@ -1,5 +1,5 @@
 import {globalState} from "./globalstate";
-import {objectAssign, deprecated, once, Lambda} from "../utils/utils";
+import {objectAssign, once, Lambda} from "../utils/utils";
 
 export function isSpyEnabled() {
 	return !!globalState.spyListeners.length;
@@ -35,17 +35,4 @@ export function spy(listener: (change: any) => void): Lambda {
 		if (idx !== -1)
 			globalState.spyListeners.splice(idx, 1);
 	});
-}
-
-export function trackTransitions(onReport?: (c: any) => void): Lambda {
-	deprecated("trackTransitions is deprecated. Use mobx.spy instead");
-	if (typeof onReport === "boolean") {
-		deprecated("trackTransitions only takes a single callback function. If you are using the mobx-react-devtools, please update them first");
-		onReport = arguments[1];
-	}
-	if (!onReport) {
-		deprecated("trackTransitions without callback has been deprecated and is a no-op now. If you are using the mobx-react-devtools, please update them first");
-		return () => {};
-	}
-	return spy(onReport);
 }