+++ src/types/observableobject.ts
@@ -10,7 +10,7 @@ import {isSpyEnabled, spyReportStart, spyReportEnd} from "../core/spy";
 
 const COMPUTED_FUNC_DEPRECATED = (
 `
-In MobX 2.* passing a function without arguments to (extend)observable will automatically be inferred to be a computed value.
+In MobX 2.* / MobX 3.* passing a function without arguments to (extend)observable will automatically be inferred to be a computed value.
 This behavior is ambiguous and will change in the future to create just an observable reference to the value passed in.
 To disambiguate, please pass the function wrapped with a modifier: use 'computed(fn)' (current behavior), 'asReference(fn)' (future behavior) or 'action(fn)'.
 Note that computed properties can also be written as '{ get propertyName() { ... }}'.